requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=5.4.1
tqdm>=4.62.0
scikit-learn>=0.24.2 